WIP: Bzlmod part the twenty-first
Thought of a way to always generate _one_ testing toolchain, but
configured to include only the required testing libraries.

Set toolchains.scalatest to default to False, so that it's only included
in extensions with `dev_dependency = True`, and must be explicitly
specified at that.

Removed stale `print()` statements in toolchain_deps.bzl.