Fix test WORKSPACE files for rules_proto 6.0.2
More to the point, now that com_google_protobuf is decoupled from
rules_proto 6, adds the necessary protobuf_deps() calls in the right
places.

I've come to realize that I don't need to update all the dependencies
before implementing Bzlmod. There's a module available for
rules_proto-5.3.0-21.7 that we can use. Then updating to rules_proto
6.0.2 can be its own separate change.

So now I'll try to get in all the Bazel 6 compatible fixes for Bazel 7,
and build Bzlmod support on top of that. And truth be told, I think the
Bzlmod support (including toolchainization) might be largely orthogonal
to those other changes anyway.