WIP: Bzlmod part the seventy-sixth
Got `./test_coverage.sh` passing by setting
`--repo_env="SCALA_VERSION=..."` in each of the test files it runs.

Set `SCALA_VERSION` to 2.12.19 in each of these files. Using other Scala
versions works, but the output is slightly different, causing the `diff`
commands in the test cases to fail.

Minor updates to turn strings in to Labels() in scala_repositories.bzl.

MODULE.bazel.lock is noisy, as always. Will probably be `.gitignore`ing
it in the end.