WIP: Bzlmod part the one hundred-eleventh
Extracted the `_get_root_module()` helper and refactored
`_get_scala_compiler_srcjars()` not to assume that module_ctx.modules[0]
is the root. `dt_patches/dt_patch_test.sh` passes.

Can probably use the result of `_get_root_module()` to pass through to
a couple of functions instead of recomputing it, but I'll tighten that
up later.