Bump Scalatest, Scalafmt in create_repository.py
Ran at the suggestion of @WojciechMazur in #1631.

After running the script, I selected these specific changes with `git
add -p`. I was pleased that the script left the Scalameta deps at
version 4.9.9, and it bumped the Scalatest versions appropriately.

However, there were a number of downgrades and updates to the Scala 2
`deps` in the Scala 3 files that I didn't accept. I'll include them in
the next commit on this branch, but do _not_ merge them.