Bump Scalatest, Scalafmt in create_repository.py
Updates Scalatest to 3.2.19 and Scalafmt to 3.8.3. Part of #1482.

After running the script at the suggestion of @WojciechMazur in #1631, I
selected these specific changes with `git add -p`. I was pleased that
the script left the Scalameta deps at version 4.9.9, and it bumped the
Scalatest versions appropriately.

However, there were a number of downgrades and updates to the Scala 2
`deps` in the Scala 3 files that I didn't accept:

- protobuf-java: 4.28.2 to 3.8.0 or 3.7.0
- scalap: from latest Scala version to an earlier version, including
  2.13.14 to 2.13.11
- com.lihaoyi.pprint: from pprint_3:0.9.0 to pprint_2.13:0.6.4
- compiler-interface: 1.10.1 to 1.3.5 or 1.9.6
- com.lihaoyi.geny: from geny_3:1.1.1 to geny_2.13:0.6.5

Plus, all the `@io_bazel_rules_scala_scala_*_2` deps explicitly added to
the `scala_3_{1,2,3,4,5}.bzl` files in #1631 for Scalafmt were stripped
of the `_2` suffix.

Perhaps it's worth updating `create_repository.py` at some point so
versions aren't automatically downgraded and `_2` suffixes are honored.