Give scala_toolchains_repo a default name argument
This turns out to be helpful when adapting
`test_version/WORKSPACE.template` to the toolchainized version of
`twitter_scrooge` for testing. In this case, we want `twitter_scooge` to
be in its own customized repo, separate from the one generated by
`scala_toolchains`.

This seems like it might be generally useful when writing module
extensions for alternative toolchains.