Third create_repository.py refactoring
Preserves the behavior of the previous version in preparation for
further improvements and eventual behavior changes. Part of #1482.

This vastly improves performance by _not_ downloading and hashing
artifacts that are already present in the current configuration.

Tested by running the original script, calling `git stash
third_party/...`, then running the updated script. Confirmed that the
updated script produced no new changes in the working tree compared to
the indexed changes.