WIP: Bzlmod part the one hundred twenty-fourth
Major revamp to extract `*_artifact_ids()` macros, and using them to
pool artifact IDs in `scala_toolchains`. This way, `scala_toolchains`
calls `repositories` with all the IDs at once, avoiding the need to
orchestrate which other macros instantiate which repos. Specifically,
the `scala_proto_instantiated` parameter for `twitter_scrooge` and
`scalafmt_repositories` is now gone.

It can be refactored a bit to be even more compact, but the proof of
concept already works under Bzlmod and `WORKSPACE`.

Also:

- Removed `scala/artifacts.bzl` and `versioned_artifact_repos`, replaced
  by `scala/scala_cross_version.bzl` and `repositories`

- Removed some `bzlmod_enabled` parameters, and repurposed the
  `twitter_scrooge` one to `register_toolchains`

- Added a `twitter_scrooge` tag class to the `deps.bzl` module
  extension, though I'm not sure if it'll ever get used

- Removed `_declare_and_bind` from `twitter_scrooge.bzl`