Compare Scala versions of `MavenCoordinates`
This is to ensure that artifacts that have a newer Scala version encoded
in the artifact name are considered newer. This takes precedence over
whether the artifact version numbers match.

Extracted `__compare_versions` from `is_newer_than` to reuse the same
code for comparing Scala versions and artifact versions. Properly
reversed the use of `lhs` and `rhs` in the process.