Enable precompiled protoc toolchain in all repos
This required importing `@bazel_features` and setting it up so we could
access `@com_google_protobuf//bazel/common/proto_common.bzl`.

Began adding documentation. Haven't tried with Bzlmod yet, but we may
need a way to create targets for users to call `register_toolchains` on.
That, or just expose `@rules_scala_toolchains` as part of the API, which
I've been avoiding.