Generate toolchains root package, update README
Now that the default Scala toolchain is optional,
`@rules_scala_toolchains()` may not contain any subpackages. This
enables `register_toolchains("@rules_scala_toolchains//...:all")` to
succeed nonetheless. This is important for the Bzlmod implementation, as
the `MODULE.bazel` file from `rules_scala` will always make this
`register_toolchains()` call.

Also updated the `README` to show directly how to apply the protocol
compiler toolchainization patch to `protobuf`. This ensures that if we
remove the patch from `scala/deps.bzl`, the example will live on in the
`README`.