Expand `$(location)` and "Make variables" in `env`
Implements the "values are subject to `$(location)` and "Make variable"
substitution" contract for the common `env` attribute for binary and
test rules. Fixes #1770.

---

Makes `env` attribute semantics for Scala binary and test targets
conformant with those documented in the Bazel docs:

- https://bazel.build/reference/be/common-definitions#common-attributes-binaries
- https://bazel.build/reference/be/common-definitions#common-attributes-tests
- https://bazel.build/reference/be/make-variables#use
- https://bazel.build/rules/lib/providers/RunEnvironmentInfo

Most of the new functions could live in a separate repository at some
point.

As for why we're using the "deprecated" `ctx.expand_make_variables`:

- https://github.com/bazelbuild/bazel/issues/5859
- https://github.com/bazelbuild/bazel-skylib/pull/486

The "deprecated" comment in the `ctx.expand_make_variables` docstring
has existed from the beginning of the file's existence (2018-05-11):

- https://github.com/bazelbuild/bazel/commit/abbb9002c41bbd53588e7249756aab236f6fcb4b